Move chat_recommendation to service
